+++ GHEtool/test/unit-tests/test_main_class.py
@@ -10,6 +10,7 @@ from GHEtool import GroundConstantTemperature, GroundFluxTemperature, FluidData,
 from GHEtool.logger import ghe_logger
 from GHEtool.Validation.cases import load_case
 from GHEtool.VariableClasses.LoadData import MonthlyGeothermalLoadAbsolute, HourlyGeothermalLoad
+from GHEtool.VariableClasses.BaseClass import UnsolvableDueToTemperatureGradient
 
 data = GroundConstantTemperature(3, 10)
 ground_data_constant = data
@@ -271,12 +272,12 @@ def test_set_pipe_params():
 
 def test_set_max_temp():
     borefield = Borefield()
-    borefield.set_max_ground_temperature(13)
+    borefield.set_max_avg_fluid_temperature(13)
     assert borefield.Tf_max == 13
-    borefield.set_max_ground_temperature(14)
+    borefield.set_max_avg_fluid_temperature(14)
     assert borefield.Tf_max == 14
     try:
-        borefield.set_max_ground_temperature(borefield.Tf_min-1)
+        borefield.set_max_avg_fluid_temperature(borefield.Tf_min - 1)
         assert False  # pragma: no cover
     except ValueError:
         assert True
@@ -284,12 +285,12 @@ def test_set_max_temp():
 
 def test_set_min_temp():
     borefield = Borefield()
-    borefield.set_min_ground_temperature(3)
+    borefield.set_min_avg_fluid_temperature(3)
     assert borefield.Tf_min == 3
-    borefield.set_min_ground_temperature(4)
+    borefield.set_min_avg_fluid_temperature(4)
     assert borefield.Tf_min == 4
     try:
-        borefield.set_min_ground_temperature(borefield.Tf_max+1)
+        borefield.set_min_avg_fluid_temperature(borefield.Tf_max + 1)
         assert False  # pragma: no cover
     except ValueError:
         assert True
@@ -308,7 +309,7 @@ def test_Tg():
 @pytest.mark.parametrize("ground_data, constant_Rb, result",
                          zip([ground_data_constant, data_ground_flux, ground_data_constant, data_ground_flux],
                              [True, True, False, False],
-                             [39.994203323480214, 38.733665022510635, 30.924434615896764, 30.245606119498383]))
+                             [39.994203323480214, 38.70946566704161, 30.924434615896764, 30.245606119498383]))
 def test_Ahmadfard(ground_data, constant_Rb, result):
     borefield = Borefield()
     borefield.borefield = copy.deepcopy(borefield_gt)
@@ -316,7 +317,7 @@ def test_Ahmadfard(ground_data, constant_Rb, result):
     borefield.set_ground_parameters(ground_data)
     borefield.set_fluid_parameters(fluidData)
     borefield.set_pipe_parameters(pipeData)
-    borefield.sizing_setup(use_constant_Rb=constant_Rb)
+    borefield.calculation_setup(use_constant_Rb=constant_Rb)
     th, qh, qm, qa = load._calculate_last_year_params(True)
     assert np.isclose(result, borefield._Ahmadfard(th, qh, qm, qa))
     assert np.isclose(result, borefield.H)
@@ -333,7 +334,7 @@ def test_Carcel(ground_data, constant_Rb, result):
     borefield.set_ground_parameters(ground_data)
     borefield.set_fluid_parameters(fluidData)
     borefield.set_pipe_parameters(pipeData)
-    borefield.sizing_setup(use_constant_Rb=constant_Rb)
+    borefield.calculation_setup(use_constant_Rb=constant_Rb)
     th, _, tcm, qh, qpm, qm = load._calculate_first_year_params(True)
     assert np.isclose(result, borefield._Carcel(th, tcm, qh, qpm, qm))
     assert np.isclose(result, borefield.H)
@@ -341,16 +342,22 @@ def test_Carcel(ground_data, constant_Rb, result):
 
 def test_set_sizing_setup():
     borefield = Borefield()
-    sizing_setup_backup = copy.deepcopy(borefield._sizing_setup)
-    borefield.sizing_setup()
-    assert borefield._sizing_setup == sizing_setup_backup
-    # set sizing_setup
+    sizing_setup_backup = copy.deepcopy(borefield._calculation_setup)
+    borefield.calculation_setup()
+    assert borefield._calculation_setup == sizing_setup_backup
+    # set calculation_setup
     test = CalculationSetup(4, False, True, False)
     test2 = CalculationSetup(3, False, False, True)
-    borefield.sizing_setup(use_constant_Rb=True, quadrant_sizing=4, L2_sizing=False, L3_sizing=True, L4_sizing=False)
-    assert borefield._sizing_setup == test
-    borefield.sizing_setup(sizing_setup=test2)
-    assert borefield._sizing_setup == test2
+    borefield.calculation_setup(use_constant_Rb=True, quadrant_sizing=4, L2_sizing=False, L3_sizing=True, L4_sizing=False)
+    assert borefield._calculation_setup == test
+    borefield.calculation_setup(calculation_setup=test2)
+    assert borefield._calculation_setup == test2
+
+    borefield.calculation_setup(rtol=10)
+    assert borefield._calculation_setup.rtol == 10
+    borefield.calculation_setup(atol=10)
+    assert borefield._calculation_setup.atol == 10
+    assert borefield._calculation_setup.rtol == 10
 
 
 def test_size():
@@ -364,9 +371,9 @@ def test_size():
     borefield.load = MonthlyGeothermalLoadAbsolute(*load_case(3))
 
     borefield.set_ground_parameters(ground_data_constant)
-    sizing_setup_backup = copy.deepcopy(borefield._sizing_setup)
+    sizing_setup_backup = copy.deepcopy(borefield._calculation_setup)
     borefield.size(L3_sizing=True)
-    assert borefield._sizing_setup == sizing_setup_backup
+    assert borefield._calculation_setup == sizing_setup_backup
 
 
 def test_select_size():
@@ -383,11 +390,11 @@ def test_select_size():
     borefield.set_ground_parameters(data_ground_flux)
     assert borefield._select_size(100, 20) == 100
     assert borefield._select_size(10, 80) == 80
-    borefield.set_max_ground_temperature(14)
+    borefield.set_max_avg_fluid_temperature(14)
     try:
         borefield._select_size(10, 80)
         assert False  # pragma: no cover
-    except ValueError:
+    except UnsolvableDueToTemperatureGradient:
         assert True
 
 
@@ -412,7 +419,7 @@ def test_size_L2_value_errors():
 
 
 @pytest.mark.parametrize("quadrant, result",
-                         zip([1, 2, 3, 4], [74.55862437702756, 96.85342542746277, 27.2041541800546, 21.789526435986094]))
+                         zip([1, 2, 3, 4], [74.55862437702756, 96.85342542746277, 27.2041541800546, 21.903857780936665]))
 def test_size_L2(quadrant, result):
     borefield = Borefield()
     borefield.borefield = copy.deepcopy(borefield_gt)
@@ -441,11 +448,11 @@ def test_size_L3_value_errors():
         assert True
 
 @pytest.mark.parametrize("quadrant, result",
-                         zip([1, 2, 3, 4], [56.37136629360852, 71.42698877336204, 26.722846792067735, 21.379839204879715]))
+                         zip([1, 2, 3, 4], [56.37136629360852, 71.42698877336204, 26.722846792067735, 21.333161686968708]))
 def test_size_L3(quadrant, result):
     borefield = Borefield()
     borefield.borefield = copy.deepcopy(borefield_gt)
-    borefield.set_max_ground_temperature(18)
+    borefield.set_max_avg_fluid_temperature(18)
     borefield.load = MonthlyGeothermalLoadAbsolute(*load_case(2))
     borefield.set_ground_parameters(ground_data_constant)
 
@@ -498,7 +505,7 @@ def test_size_L4():
     load.load_hourly_profile(FOLDER.joinpath("Examples/hourly_profile.csv"), col_cooling=0, col_heating=1)
     borefield.load = load
 
-    borefield.set_max_ground_temperature(25)
+    borefield.set_max_avg_fluid_temperature(25)
     assert np.isclose(109.4742962707615, borefield.size_L4(100, quadrant_sizing=3))
     assert np.isclose(109.4742962707615, borefield.H)
     assert borefield.calculate_quadrant() == 3
@@ -508,7 +515,7 @@ def test_size_L4():
     borefield.load = load
 
     # to increase coverage
-    borefield.sizing_setup(L4_sizing=True)
+    borefield.calculation_setup(L4_sizing=True)
     assert np.isclose(174.23648328808213, borefield.size(100, quadrant_sizing=4))
     assert np.isclose(174.23648328808213, borefield.H)
     assert borefield.calculate_quadrant() == 4
@@ -570,7 +577,7 @@ def test_calculate_temperatures():
     borefield.load = MonthlyGeothermalLoadAbsolute(*load_case(2))
 
     borefield.calculate_temperatures(120)
-    np.testing.assert_array_almost_equal(borefield.results_peak_heating,
+    np.testing.assert_array_almost_equal(borefield.results.peak_heating,
                                          np.array([ 7.76418446,  8.06457527,  8.59144992,  9.22406634,  9.9176661 ,
        10.35493277, 10.85933332, 10.99023511, 10.04321457,  9.33984947,
         8.73958309,  8.33426967,  7.95321315,  8.25111367,  8.77481226,
@@ -619,7 +626,7 @@ def test_calculate_temperatures():
        10.55323455,  9.93631861,  9.51283391,  9.1158352 ,  9.39905299,
         9.90536243, 10.51952398, 11.19933578, 11.62696168, 12.12650419,
        12.25159304, 11.29457142, 10.57776036,  9.96084442,  9.53735972]))
-    np.testing.assert_array_almost_equal(borefield.results_peak_cooling,
+    np.testing.assert_array_almost_equal(borefield.results.peak_cooling,
                                          np.array([ 9.7265228 ,  9.87217467, 10.134688  , 10.63927884, 11.62475183,
        12.59873554, 13.28921831, 13.76520367, 12.39881805, 10.70464274,
        10.2834413 , 10.00058989,  9.91555149, 10.05871306, 10.31805035,
@@ -668,7 +675,7 @@ def test_calculate_temperatures():
        11.91802782, 11.48017683, 11.17915413, 11.07817353, 11.20665239,
        11.44860052, 11.93473647, 12.90642151, 13.87076446, 14.55638918,
        15.0265616 , 13.6501749 , 11.94255363, 11.50470264, 11.20367994]))
-    np.testing.assert_array_almost_equal(borefield.results_month_cooling,
+    np.testing.assert_array_almost_equal(borefield.results.monthly_cooling,
                                          np.array([ 9.7265228 ,  9.87217467,  9.91023486,  9.96749514, 10.24004053,
        10.62890537, 11.40727853, 11.53818031, 10.77407145, 10.54689525,
        10.2834413 , 10.00058989,  9.91555149, 10.05871306, 10.09359721,
@@ -717,7 +724,7 @@ def test_calculate_temperatures():
        11.76028033, 11.48017683, 11.17915413, 11.07817353, 11.20665239,
        11.22414738, 11.26295278, 11.52171021, 11.90093428, 12.67444939,
        12.79953825, 12.0254283 , 11.78480614, 11.50470264, 11.20367994]))
-    np.testing.assert_array_almost_equal(borefield.results_month_heating,
+    np.testing.assert_array_almost_equal(borefield.results.monthly_heating,
                                          np.array([ 9.37492462,  9.46486873,  9.54493806,  9.64968692,  9.9176661 ,
        10.35493277, 10.85933332, 10.99023511, 10.38868332, 10.18251169,
         9.93275637,  9.66908304,  9.56395331,  9.65140713,  9.72830041,
@@ -783,14 +790,14 @@ def test_gfunction():
     try:
         borefield.gfunction(56491)
         assert False  # pragma: no cover
-    except ValueError:
+    except UnsolvableDueToTemperatureGradient:
         assert True
     borefield.H = 102.3
 
     np.testing.assert_array_almost_equal(borefield.gfunction([6000, 60000, 600000]), np.array([0.63751082, 1.70657847, 2.84227252]))
     borefield.create_custom_dataset()
     np.testing.assert_array_almost_equal(borefield.gfunction([6000, 60000, 600000]), np.array([0.622017, 1.703272, 2.840246]))
-    borefield.sizing_setup(use_precalculated_dataset=False)
+    borefield.calculation_setup(use_precalculated_dataset=False)
     np.testing.assert_array_almost_equal(borefield.gfunction([6000, 60000, 600000]), np.array([0.63751082, 1.70657847, 2.84227252]))
 
 
@@ -809,7 +816,7 @@ def test_load_duration(monkeypatch):
 def test_calculate_quadrants_without_data():
     borefield = Borefield()
     borefield.borefield = copy.deepcopy(borefield_gt)
-    borefield.set_max_ground_temperature(18)
+    borefield.set_max_avg_fluid_temperature(18)
     borefield.load = MonthlyGeothermalLoadAbsolute(*load_case(2))
     borefield.set_ground_parameters(ground_data_constant)
     borefield.calculate_quadrant()
@@ -874,7 +881,7 @@ def test_effect_H_init(H, result):
     borefield.create_rectangular_borefield(10, 5, 7, 7, 100, 0.75)
     load = MonthlyGeothermalLoadAbsolute(*load_case(1))
     borefield.load = load
-    borefield.sizing_setup(H_init=H)
+    borefield.calculation_setup(H_init=H)
     assert np.isclose(borefield.size(), result)
     borefield = Borefield()
     borefield.ground_data = GroundConstantTemperature(3, 11)
@@ -886,7 +893,7 @@ def test_effect_H_init(H, result):
 
 def test_depth_convergence():
     borefield = Borefield()
-    borefield.sizing_setup(atol=1, rtol=0.01, max_nb_of_iterations=10)
+    borefield.calculation_setup(atol=1, rtol=0.01, max_nb_of_iterations=10)
     try:
         borefield._check_convergence(10, 12, 10)
         assert False  # pragma: no cover
@@ -897,3 +904,40 @@ def test_depth_convergence():
     assert not borefield._check_convergence(10, 10.5, 1)
     assert borefield._check_convergence(10, 10.001, 1)
     assert not borefield._check_convergence(10000, 10002, 1)
+
+
+def test_calculate_next_depth_deep_sizing():
+    borefield = Borefield()
+    borefield.ground_data = GroundFluxTemperature(3, 10)
+    borefield.create_rectangular_borefield(10, 5, 7, 7, 100, 0.75)
+    load = MonthlyGeothermalLoadAbsolute(*load_case(1))
+    borefield.load = load
+
+    borefield.calculate_temperatures(75)
+    assert np.isclose(borefield.calculate_next_depth_deep_sizing(75), 117.98660599828808)
+    borefield.calculate_temperatures(117.98660599828808)
+    assert np.isclose(borefield.calculate_next_depth_deep_sizing(117.98660599828808), 128.16618036528823)
+    borefield.calculate_temperatures(128.16618036528823)
+    assert np.isclose(borefield.calculate_next_depth_deep_sizing(128.16618036528823), 130.8812255630479)
+
+
+@pytest.mark.parametrize("case, result",
+                         zip((1, 2, 3, 4),
+                             [131.90418292004594, 0, 139.46239300837794, 131.90418292004594]))
+def test_deep_sizing(case, result):
+    borefield = Borefield()
+    borefield.ground_data = GroundFluxTemperature(3, 10)
+    borefield.create_rectangular_borefield(10, 5, 7, 7, 100, 0.75)
+    load = MonthlyGeothermalLoadAbsolute(*load_case(case))
+    borefield.load = load
+
+    assert np.allclose(result, borefield._size_based_on_temperature_profile(10, deep_sizing=True)[0])
+
+    borefield = Borefield()
+    borefield.ground_data = GroundFluxTemperature(3, 10)
+    borefield.create_rectangular_borefield(10, 5, 7, 7, 100, 0.75)
+    load = MonthlyGeothermalLoadAbsolute(*load_case(case))
+    borefield.load = load
+
+    # methods should more or less lead to the same results, 1% diff taken as a reference
+    assert np.allclose(result, borefield._size_based_on_temperature_profile(10, deep_sizing=False)[0], rtol=0.01)