+++ CHANGELOG.md
@@ -11,9 +11,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - New class of GroundData: GroundTemperatureGradient added (issue #145).
 - Load classes (issue #45).
 - Pipe classes (single, double, coaxial, Multiple U Tube) (issue #40 and #45).
+- Added another methodology for sizing with a variable ground temperature (issue #144).
+- Custom error when the field cannot be sized due to a ground temperature gradient (issue #156).
 - Interpolation option in calculate function in Gfunction class (issue #159).
 - Absolute and relative tolerances for the sizing methods even as a maximum number of iterations is added, so there is more transparency and flexibility in the trade-off between accuracy and speed (issue #159).
 - Added advanced options to GHEtool GUI (issue #165).
+- Added a result class so all calculated temperatures are now in a separate Result class object within the borefield object (issue #167).
 - Added domestic hot water (DHW) to GHEtool (issue #172).
 
 ## Changed
@@ -24,10 +27,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Removed 'set_hourly_cooling_load', 'set_hourly_heating_load' from main_class and move it to separate load class (issue #45).
 - Moved draw_borehole_internals to PipeClass (issue #45).
 - Borehole equivalent resistances is now calculated in one step, centralised in the pipe class (issue #45).
-- Go to 100% code coverage with 300 tests.
+- Go to 100% code coverage with 350 tests.
+- Threshold interpolation for g-functions set to a relative threshold of 25% relative to the demanded depth (issue #144).
+- Implemented a custom error for crossing the maximum number of iterations: 'MaximumNumberOfIterations' (issue #144).
+- _size_based_on_temperature_profile now returns two arguments: the required depth and a boolean flag to check if the field is sized (issue #144).
+- Speed up of L3/L4 sizing by halving calculation time due to intermediate checks if the field is sized (issue #144).
+- Changed ValueError when the field cannot be sized due to a temperature gradient to the custom UnsolvableDueToTemperatureGradient Exception (issue #156).
 - Rename SizingSetup class to CalculationSetup class (issue #159).
 - Move H_init to CalculationSetup class (issue #159).
 - Move use_precalcated_data to CalculationSetup class and rename to: 'use_precalculate_dataset' (issue #159).
+- Changed 'set_max_ground_temperature' and 'set_min_ground_temperature' to correct names: 'set_max_avg_fluid_temperature' and 'set_min_avg_fluid_temperature'
 - Changed 'minimal average fluid temperature' to 'minimum average fluid temperature' in GUI (issue #172).
 
 ## Fixed